Fixes as per comments.
